Design and mobile layout changes
Html:  Only the container div was moved so that it was no longer effecting the sizing of the main and screen.
JS: I only altered a few things to test some thing out, but put them back the way they were.
CSS: added some sizing, and design things, as well as a some sizing to make the layout mobile friendly.